Add initial scroll position hook to HTML spec
scroll-start-target is a CSS property which affects an element's
initial scroll position. A scroll container's initial scroll position
determines what scroll offset that scroll container should be at before
any "explicit" scrolling (e.g. user gesture, programmatic scroll API)
occurs on it. This initial scroll position is affected by CSS
properties such as scroll-start-target and might change as a document
is loaded.

This patch adds a step in the "Updating the  document" section of the
HTML spec to direct user agents on when to re-evaluate initial scroll
positions and adjust accordingly.